requirements.txt
matplotlib==3.6.0
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.2.0
streamlit==1.13.0
torch==1.12.1






